Add structure for creating demo content
Add a feature 'os_project_demo_content' which uses UUID and deploy to export
content to a feature which can then be imported simply by enabling the feature.

Not sure the extent to which this can be used yet, but at the very least it
allows the exporting of basic pages, which should speed up the process of
adding demo content. By default, the feature won't work because the extra make
file, 'demo.make' is disabled by default, which provides the necessary
dependencies.